requirements.txt
python-dotenv
requests
pyTelegramBotAPI
fpdf

